backend/requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
alembic==1.17.0
psycopg2-binary==2.9.11
redis==7.0.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
openai==2.6.0
pydantic-settings==2.11.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.185.0
cryptography==46.0.3
bcrypt==5.0.0
email-validator==2.3.0
isodate==0.7.2
